+++ packages/radio-player/index.js
@@ -7,8 +7,8 @@ import { AudioSource } from '@internetarchive/audio-element';
 import RadioPlayer from './lib/src/radio-player';
 import RadioPlayerConfig from './lib/src/models/radio-player-config';
 import { SearchHandler } from './lib/src/search-handler/search-handler';
-import { LocalSearchIndex } from './lib/src/search-handler/search-indices/local-search-index';
-import { RadioArchiveSearchIndex } from './lib/src/search-handler/search-indices/radio-archive-search-index';
+import { LocalSearchBackend } from './lib/src/search-handler/search-backends/local-search-backend/local-search-backend';
+import { FullTextSearchBackend } from './lib/src/search-handler/search-backends/full-text-search-backend/full-text-search-backend';
 import { TranscriptIndex } from './lib/src/search-handler/transcript-index';
 
 export {
@@ -18,7 +18,7 @@ export {
   TranscriptEntryConfig,
   AudioSource,
   SearchHandler,
-  LocalSearchIndex,
-  RadioArchiveSearchIndex,
+  LocalSearchBackend,
+  FullTextSearchBackend,
   TranscriptIndex
 };
