+++ packages/radio-player/test/search-handler/search-handler.test.js
@@ -4,7 +4,7 @@ import {
 
 import { TranscriptConfig, TranscriptEntryConfig } from "@internetarchive/transcript-view";
 import { SearchHandler } from '../../lib/src/search-handler/search-handler';
-import { LocalSearchIndex } from '../../lib/src/search-handler/search-indices/local-search-index';
+import { LocalSearchBackend } from '../../lib/src/search-handler/search-backends/local-search-backend/local-search-backend';
 import { TranscriptIndex } from '../../lib/src/search-handler/transcript-index';
 
 class MockSearchIndex {
@@ -39,7 +39,7 @@ describe('Search Handler', () => {
       const entry3 = new TranscriptEntryConfig(3, 10, 13, 'bump baz boing', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const response = searchHandler.getSearchSeparatedTranscript('ba');
@@ -53,7 +53,7 @@ describe('Search Handler', () => {
       const entry3 = new TranscriptEntryConfig(3, 10, 13, 'bump baz boing', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const response = searchHandler.getSearchSeparatedTranscript('zip');
@@ -67,7 +67,7 @@ describe('Search Handler', () => {
       const entry3 = new TranscriptEntryConfig(3, 10, 13, 'bump baz boing', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const results = await searchHandler.getSearchSeparatedTranscript('zip');
@@ -87,7 +87,7 @@ describe('Search Handler', () => {
       const entry3 = new TranscriptEntryConfig(3, 10, 13, 'bump baz boing', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const transcriptSearchResults = await searchHandler.getSearchSeparatedTranscript('ba');
@@ -146,7 +146,7 @@ describe('Search Handler', () => {
       const entry3 = new TranscriptEntryConfig(3, 10, 13, 'bump boing', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('boop');
@@ -190,7 +190,7 @@ describe('Search Handler', () => {
       const entry5 = new TranscriptEntryConfig(5, 19, 23, 'grip grap', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4, entry5]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('bump');
@@ -229,7 +229,7 @@ describe('Search Handler', () => {
       const entry7 = new TranscriptEntryConfig(7, 29, 34, 'ooga booga', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4, entry5, entry6, entry7]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('bump');
@@ -258,7 +258,7 @@ describe('Search Handler', () => {
       const entry4 = new TranscriptEntryConfig(4, 14, 18, 'fizz buzz', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('blop bump');
@@ -274,7 +274,7 @@ describe('Search Handler', () => {
       const entry5 = new TranscriptEntryConfig(5, 19, 23, 'snip snap', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4, entry5]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('blop bump boing fizz');
@@ -324,7 +324,7 @@ describe('Search Handler', () => {
       const entry4 = new TranscriptEntryConfig(4, 14, 18, 'fizz buzz', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('bu');
@@ -353,7 +353,7 @@ describe('Search Handler', () => {
       const entry4 = new TranscriptEntryConfig(4, 15.045, 20.033, 'Baltimore area district which the president called Rat and road and infested', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('the');
@@ -403,7 +403,7 @@ describe('Search Handler', () => {
       const entry4 = new TranscriptEntryConfig(4, 14, 18, 'fizz buzz', false);
       const transcriptConfig = new TranscriptConfig([entry1, entry2, entry3, entry4]);
       const transcriptIndex = new TranscriptIndex(transcriptConfig);
-      const searchIndex = new LocalSearchIndex(transcriptIndex);
+      const searchIndex = new LocalSearchBackend(transcriptIndex);
       const searchHandler = new SearchHandler(searchIndex, transcriptIndex);
 
       const newTranscript = await searchHandler.search('blop bump');
